Move degree status field to its own step
The question "What stage are you at with your degree?" is presumptive and might
be contributing towards drop outs. We want to remove the question from page 1
and then rephrase the question/answer and add to a new page 2.

Remove `degree_status_id` field from the `Name` step and add to a new
`DegreeStatus` step.

Change from a select box to a radio input.

Update copy of the answers to be clearer/more specific. If a translation
doesn't exist (for a new CRM option, for example) then it will default to the
text in the CRM.
