+++ app/views/mailing_list/steps/_degree_status.html.erb
@@ -1,0 +1,14 @@
+<h1>Do you have a degree?</h1>
+
+<p>
+Telling us whether you have a degree, or are studying for a degree, will help us to give you the right support to become a teacher.
+</p>
+
+<%= f.govuk_error_summary %>
+
+<%= f.govuk_collection_radio_buttons :degree_status_id,
+  f.object.degree_status_options,
+  :id,
+  ->(option) { I18n.t("helpers.answer.mailing_list_steps_degree_status.degree_status.#{option.value.parameterize(separator: "_")}", default: option.value) },
+  legend: { text: t('helpers.label.mailing_list_steps_degree_status.degree_status_id') }
+%>
