+++ spec/factories/mailing_list/name_factory.rb
@@ -3,6 +3,5 @@ FactoryBot.define do
     first_name { "Test" }
     sequence(:last_name) { |n| "User #{n}" }
     sequence(:email) { |n| "testuser#{n}@testing.education.gov.uk" }
-    degree_status_id { GetIntoTeachingApiClient::Constants::DEGREE_STATUS_OPTIONS["First year"] }
   end
 end
